\name{cutmondate-methods}
\docType{methods}
\alias{cutmondate}
\alias{cutmondate-methods}
\alias{cutmondate,mondate-method}
\alias{cutmondate,Date-method}
\alias{cutmondate,POSIXlt-method}
\alias{cutmondate,POSIXct-method}
\title{Methods to Generate Date Cuts}
\description{
Methods to cut sequences of mondate, Date, and POSIXt objects.
This is essentially a wrapper for the \code{cut.mondate} method.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "mondate")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "Date")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "POSIXct")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "POSIXlt")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
