\name{cutmondate-methods}
\docType{methods}
\alias{cutmondate}
\alias{cutmondate-methods}
\alias{cutmondate,mondate-method}
\alias{cutmondate,Date-method}
\alias{cutmondate,POSIXlt-method}
\alias{cutmondate,POSIXct-method}
\title{Methods to Generate Date Cuts}
\description{
Methods to cut sequences of mondate, Date, and POSIXt objects.
}
\keyword{methods}
